adding controller advice
